OoprRoxy Internals

OoprRoxy is called by roxygen2 methods. It creates OoprRoxyClass which creates a roxy block containing sections for each class.

TODO:

Try to constrain horizontal width for nested sections.

Tidy up pdf

Usage

OoprRoxy()

OoprRoxySection(title = "", content = character(0L), hr = FALSE, pfx = "")

OoprRoxyDescribe(title = "Fields", hr = FALSE)

OoprRoxyUsage(content = character(0L), name = "")

OoprRoxyArguments(args = list())

OoprRoxyMethod(title, tags, fun, warn = TRUE, hr = TRUE, pfx = "")

OoprRoxyClass(block)

Arguments

title

character(1L)
The title of the section.

content

character()
Lines of the content for the section.

hr

logical(1L)
Whether to add horizontal line to section heading.

pfx

character(1L)
To add a hyperref.

args

list()
A list of roxy_tag_params.

tags

list()
A list of (potentially optional) tags: @description, @usage, @param, @details & @returns.

fun

function
The function object of the method.

warn

logical(1L)
Whether warnings should display when missing tags.

block

roxy_block
A roxy block containing an oopr class.

Value

OoprRoxySection

[3 fields] [4 methods]
Create a subsection.

OoprRoxyDescribe

[4 fields] [3 methods]
Represents a describe subsection.

OoprRoxyUsage

[3 fields] [3 methods]
Represents a usage subsection.

OoprRoxyArguments

[4 fields] [3 methods]
Create an arguments subsection.

OoprRoxyMethod

[6 fields] [3 methods]
Represents a method subsection.

OoprRoxyClass

[5 fields] [4 methods]
Represents a class section.

OoprRoxySection

Description

Create a subsection.

Fields

title

[read-only]
character(1L)
The title of the section.

content

[read-only]
character()
Lines of the content for the section.

size

[read-only]
integer(1L)
The number of lines.

Methods

insert

[virtual]
Insert a line into the section.

erase

[final]
Remove all lines from the section.

toRd

[final]
Convert the section into Rd formatted text.

format

[protected] [virtual]
Convert $content into a format for use in $toRd().

insert

Description

Insert a line into the section.

Usage

insert(x, i = this$size + 1L)

Arguments

x	character()
The line to insert.

i	integer() | character()
The position to insert, must be the same length as x.

Returns

this invisibly.

erase

Description

Remove all lines from the section.

Usage

erase()

Returns

this invisibly.

toRd

Description

Convert the section into Rd formatted text.

Usage

toRd()

Details

Uses the output from virtual method $format().

Returns

character(1L), wrapped in \subsection.

format

Description

Convert $content into a format for use in $toRd().

Usage

format()

Returns

Must return character().

OoprRoxyDescribe

Description

Represents a describe subsection.

Fields

names

[read-only]
character(0L)
Name of each item.

title

[read-only]
character(1L)
The title of the section.

content

[read-only]
character()
Lines of the content for the section.

size

[read-only]
integer(1L)
The number of lines.

Methods

insert

[virtual]
Insert a line into the section.

erase

[final]
Remove all lines from the section.

toRd

[final]
Convert the section into Rd formatted text.

insert

Description

Insert a line into the section.

Usage

insert(x, i = character(1L))

Arguments

x	character() | roxy_tag_field
Item to insert. If a field roxy tag, then takes the $val contents to set x and i.

i	character()
The name of the item.

Returns

this invisibly.

erase

Description

Remove all lines from the section.

Usage

erase()

Returns

this invisibly.

toRd

Description

Convert the section into Rd formatted text.

Usage

toRd()

Details

Uses the output from virtual method $format().

Returns

character(1L), wrapped in \subsection.

OoprRoxyUsage

Description

Represents a usage subsection.

Fields

title

[read-only]
character(1L)
The title of the section.

content

[read-only]
character()
Lines of the content for the section.

size

[read-only]
integer(1L)
The number of lines.

Methods

insert

[virtual]
Insert a line into the section.

erase

[final]
Remove all lines from the section.

toRd

[final]
Convert the section into Rd formatted text.

insert

Description

Insert a line into the section.

Usage

insert(x, i = this$size + 1L)

Arguments

x	character()
The line to insert.

i	integer() | character()
The position to insert, must be the same length as x.

Returns

this invisibly.

erase

Description

Remove all lines from the section.

Usage

erase()

Returns

this invisibly.

toRd

Description

Convert the section into Rd formatted text.

Usage

toRd()

Details

Uses the output from virtual method $format().

Returns

character(1L), wrapped in \subsection.

OoprRoxyArguments

Description

Create an arguments subsection.

Fields

names

[read-only]
character(0L)
Name of each item.

title

[read-only]
character(1L)
The title of the section.

content

[read-only]
character()
Lines of the content for the section.

size

[read-only]
integer(1L)
The number of lines.

Methods

insert

[virtual]
Insert a line into the section.

erase

[final]
Remove all lines from the section.

toRd

[final]
Convert the section into Rd formatted text.

insert

Description

Insert a line into the section.

Usage

insert(x, i = character(1L))

Arguments

x	character() | roxy_tag_param
Item to insert. If a param roxy tag, then takes the $val contents to set x and i.

i	character()
The name of the argument.

Returns

this invisibly.

erase

Description

Remove all lines from the section.

Usage

erase()

Returns

this invisibly.

toRd

Description

Convert the section into Rd formatted text.

Usage

toRd()

Details

Uses the output from virtual method $format().

Returns

character(1L), wrapped in \subsection.

OoprRoxyMethod

Description

Represents a method subsection.

Details

Combines many sections into a single subsection. If the tags @description, @returns and @param are not provided, a warning will display.

Fields

title

[read-only]
character(1L)
The title of the section.

fun

[read-only]
function
The function object of the method.

warn

[read-only]
logical(1L)
Whether warnings should display when missing tags.

sections

[container]
OoprRoxySection
A container of the subsections.

content

[read-only]
character()
Lines of the content for the section.

size

[read-only]
integer(1L)
The number of lines.

Methods

insert

[virtual]
Insert a line into the section.

erase

[final]
Remove all lines from the section.

toRd

[final]
Convert the section into Rd formatted text.

insert

Description

Insert a line into the section.

Usage

insert(x, i = this$size + 1L)

Arguments

x	character()
The line to insert.

i	integer() | character()
The position to insert, must be the same length as x.

Returns

this invisibly.

erase

Description

Remove all lines from the section.

Usage

erase()

Returns

this invisibly.

toRd

Description

Convert the section into Rd formatted text.

Usage

toRd()

Details

Uses the output from virtual method $format().

Returns

character(1L), wrapped in \subsection.

OoprRoxyClass

Description

Represents a class section.

Details

Combines many sections into a single top-level section for a class.

Methods need to be called in order.

Fields

title

[read-only]
character(1L)
The title of the section.

block

[read-only]
roxy_block
A roxy block containing an oopr class.

members

[read-only]
list()
A (nested) list of tags for each member.

rdname

[read-only]
character(1L)
The name of the .Rd.

sections

[container]
OoprRoxySection
The subsections inside the class section.

Methods

makeSections

Creates subsections for @description and @details tags.

makeFields

Creates a subsection for a list of fields/properties.

makeMethods

Creates a subsection for a list of methods and inserts method sections.

makeTag

Places the class section under the blocks tags list

makeSections

Description

Creates subsections for @description and @details tags.

Usage

makeSections()

makeFields

Description

Creates a subsection for a list of fields/properties.

Usage

makeFields()

makeMethods

Description

Creates a subsection for a list of methods and inserts method sections.

Usage

makeMethods()

makeTag

Description

Places the class section under the blocks tags list

Usage

makeTag()